feat: data-binding demo
