+++ src/app/layout/layout.component.html
@@ -1,5 +1,5 @@
 <div class="min-height-300 bg-primary position-absolute w-100"></div>
-
+<!-- Header -->
 <aside
   class="sidenav bg-white navbar navbar-vertical navbar-expand-xs border-0 border-radius-xl my-3 fixed-start ms-4"
   id="sidenav-main"
@@ -27,11 +27,7 @@
   <div class="collapse navbar-collapse w-auto" id="sidenav-collapse-main">
     <ul class="navbar-nav">
       <li class="nav-item">
-        <a
-          class="nav-link active"
-          [routerLink]="['/index']"
-          routerLinkActive="router-link-active"
-        >
+        <a class="nav-link" [routerLink]="['/index']" routerLinkActive="active">
           <div
             class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center"
           >
@@ -41,7 +37,7 @@
         </a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" routerLink="/table">
+        <a class="nav-link" routerLink="/table" routerLinkActive="active">
           <div
             class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center"
           >
