+++ src/app/page/member/index-routing.module.ts
@@ -1,12 +1,19 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { IndexComponent } from './index.component';
+import { IndexComponent } from './index/index.component';
+import { TableComponent } from './table/table.component';
 
 /** 首頁 路由 */
 const routes: Routes = [
   {
     path: '',
-    children: [{ path: '', component: IndexComponent }],
+    children: [
+      {
+        path: 'index',
+        component: IndexComponent,
+      },
+      { path: 'table', component: TableComponent },
+    ],
   },
 ];
