--- src/app/page/table/table.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-table',
-  templateUrl: './table.component.html',
-  styleUrls: ['./table.component.scss']
-})
-export class TableComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
