+++ src/index.html
@@ -13,28 +13,57 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <link rel="apple-touch-icon" sizes="76x76" href="./assets/img/apple-icon.png" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1, shrink-to-fit=no"
+    />
+    <link
+      rel="apple-touch-icon"
+      sizes="76x76"
+      href="./assets/img/apple-icon.png"
+    />
     <link rel="icon" type="image/png" href="./assets/img/favicon.png" />
     <title>Argon Dashboard 2 by Creative Tim</title>
     <base href="/" />
     <!--     Fonts and icons     -->
-    <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet" />
+    <link
+      href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700"
+      rel="stylesheet"
+    />
     <!-- Nucleo Icons -->
     <link href="./assets/css/nucleo-icons.css" rel="stylesheet" />
     <link href="./assets/css/nucleo-svg.css" rel="stylesheet" />
     <!-- Font Awesome Icons -->
-    <script src="https://kit.fontawesome.com/42d5adcbca.js" crossorigin="anonymous"></script>
+    <script
+      src="https://kit.fontawesome.com/42d5adcbca.js"
+      crossorigin="anonymous"
+    ></script>
     <link href="./assets/css/nucleo-svg.css" rel="stylesheet" />
     <!-- CSS Files -->
-    <link id="pagestyle" href="./assets/css/argon-dashboard.css?v=2.0.4" rel="stylesheet" />
+    <link
+      id="pagestyle"
+      href="./assets/css/argon-dashboard.css?v=2.0.4"
+      rel="stylesheet"
+    />
     <link rel="preconnect" href="https://fonts.gstatic.com" />
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap" rel="stylesheet" />
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet" />
-      <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-</head>
+    <link
+      href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap"
+      rel="stylesheet"
+    />
+    <link
+      href="https://fonts.googleapis.com/icon?family=Material+Icons"
+      rel="stylesheet"
+    />
+    <link rel="preconnect" href="https://fonts.gstatic.com" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap"
+      rel="stylesheet"
+    />
+    <link
+      href="https://fonts.googleapis.com/icon?family=Material+Icons"
+      rel="stylesheet"
+    />
+  </head>
 
   <body class="g-sidenav-show bg-gray-100 mat-typography">
     <app-root></app-root>
@@ -45,24 +74,34 @@
     <script src="./assets/js/plugins/smooth-scrollbar.min.js"></script>
     <script src="./assets/js/plugins/chartjs.min.js"></script>
     <script>
-      var ctx1 = document.getElementById('chart-line').getContext('2d')
+      var ctx1 = document.getElementById("chart-line").getContext("2d");
 
-      var gradientStroke1 = ctx1.createLinearGradient(0, 230, 0, 50)
+      var gradientStroke1 = ctx1.createLinearGradient(0, 230, 0, 50);
 
-      gradientStroke1.addColorStop(1, 'rgba(94, 114, 228, 0.2)')
-      gradientStroke1.addColorStop(0.2, 'rgba(94, 114, 228, 0.0)')
-      gradientStroke1.addColorStop(0, 'rgba(94, 114, 228, 0)')
+      gradientStroke1.addColorStop(1, "rgba(94, 114, 228, 0.2)");
+      gradientStroke1.addColorStop(0.2, "rgba(94, 114, 228, 0.0)");
+      gradientStroke1.addColorStop(0, "rgba(94, 114, 228, 0)");
       new Chart(ctx1, {
-        type: 'line',
+        type: "line",
         data: {
-          labels: ['Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
+          labels: [
+            "Apr",
+            "May",
+            "Jun",
+            "Jul",
+            "Aug",
+            "Sep",
+            "Oct",
+            "Nov",
+            "Dec",
+          ],
           datasets: [
             {
-              label: 'Mobile apps',
+              label: "Mobile apps",
               tension: 0.4,
               borderWidth: 0,
               pointRadius: 0,
-              borderColor: '#5e72e4',
+              borderColor: "#5e72e4",
               backgroundColor: gradientStroke1,
               borderWidth: 3,
               fill: true,
@@ -81,7 +120,7 @@
           },
           interaction: {
             intersect: false,
-            mode: 'index',
+            mode: "index",
           },
           scales: {
             y: {
@@ -95,11 +134,11 @@
               ticks: {
                 display: true,
                 padding: 10,
-                color: '#fbfbfb',
+                color: "#fbfbfb",
                 font: {
                   size: 11,
-                  family: 'Open Sans',
-                  style: 'normal',
+                  family: "Open Sans",
+                  style: "normal",
                   lineHeight: 2,
                 },
               },
@@ -114,27 +153,27 @@
               },
               ticks: {
                 display: true,
-                color: '#ccc',
+                color: "#ccc",
                 padding: 20,
                 font: {
                   size: 11,
-                  family: 'Open Sans',
-                  style: 'normal',
+                  family: "Open Sans",
+                  style: "normal",
                   lineHeight: 2,
                 },
               },
             },
           },
         },
-      })
+      });
     </script>
     <script>
-      var win = navigator.platform.indexOf('Win') > -1
-      if (win && document.querySelector('#sidenav-scrollbar')) {
+      var win = navigator.platform.indexOf("Win") > -1;
+      if (win && document.querySelector("#sidenav-scrollbar")) {
         var options = {
-          damping: '0.5',
-        }
-        Scrollbar.init(document.querySelector('#sidenav-scrollbar'), options)
+          damping: "0.5",
+        };
+        Scrollbar.init(document.querySelector("#sidenav-scrollbar"), options);
       }
     </script>
     <!-- Github buttons -->
@@ -144,4 +183,3 @@
     <app-root></app-root>
   </body>
 </html>
-
